requirements.txt
duckdb>=1.0.0
pandas>=2.2.0
requests>=2.32.0
typer>=0.12.0
pydantic>=2.7.0
PyYAML>=6.0.0
jinja2>=3.1.0
plotly>=5.22.0
tenacity>=8.3.0
pyarrow>=14.0.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
pyjwt>=2.8.0
bcrypt>=4.1.0